Izmene vezane za use case i class dijagrame

+++ Projekat/bolnica/bolnica/Forms/FormUpravnik.xaml.cs
@@ -159,47 +159,27 @@ namespace bolnica.Forms
 
         private void Button_Click_Obrisi(object sender, RoutedEventArgs e)
         {
-            Prostorija row = (Prostorija)dataGridProstorije.SelectedItems[0];
-            if (row.Zauzeta)
+            if (dataGridProstorije.SelectedCells.Count > 0)
             {
-                MessageBox.Show("Prostorija je trenutno zauzeta, ne možete je obrisati.");
-            }
-            else
-            {
-                List<Prostorija> prostorije = storage.GetAllProstorije();
-                List<BolnickaSoba> bolnickeSobe = storage.GetAllBolnickeSobe();
-
-                var s = new CreateFormProstorije();
-                bool found = false;
-                foreach (Prostorija p in prostorije)
+                Prostorija row = (Prostorija)dataGridProstorije.SelectedItems[0];
+                if (row.Zauzeta)
                 {
-                    if (p.BrojProstorije == row.BrojProstorije)
-                    {
-                        storage.Delete(p);
-                        p.Obrisana = true;
-                        storage.Save(p);
-
-                        for (int i = 0; i < FormUpravnik.Prostorije.Count; i++)
-                        {
-                            if (FormUpravnik.Prostorije[i].BrojProstorije == row.BrojProstorije)
-                            {
-                                FormUpravnik.Prostorije.Remove(FormUpravnik.Prostorije[i]);
-                            }
-                        }
-                        found = true;
-                        break;
-                    }
+                    MessageBox.Show("Prostorija je trenutno zauzeta, ne možete je obrisati.");
                 }
-
-                if (!found)
+                else
                 {
-                    foreach (BolnickaSoba b in bolnickeSobe)
+                    List<Prostorija> prostorije = storage.GetAllProstorije();
+                    List<BolnickaSoba> bolnickeSobe = storage.GetAllBolnickeSobe();
+
+                    var s = new CreateFormProstorije();
+                    bool found = false;
+                    foreach (Prostorija p in prostorije)
                     {
-                        if (b.BrojProstorije == row.BrojProstorije)
+                        if (p.BrojProstorije == row.BrojProstorije)
                         {
-                            storage.Delete(b);
-                            b.Obrisana = true;
-                            storage.Save(b);
+                            storage.Delete(p);
+                            p.Obrisana = true;
+                            storage.Save(p);
 
                             for (int i = 0; i < FormUpravnik.Prostorije.Count; i++)
                             {
@@ -208,8 +188,31 @@ namespace bolnica.Forms
                                     FormUpravnik.Prostorije.Remove(FormUpravnik.Prostorije[i]);
                                 }
                             }
-
+                            found = true;
                             break;
+                        }
+                    }
+
+                    if (!found)
+                    {
+                        foreach (BolnickaSoba b in bolnickeSobe)
+                        {
+                            if (b.BrojProstorije == row.BrojProstorije)
+                            {
+                                storage.Delete(b);
+                                b.Obrisana = true;
+                                storage.Save(b);
+
+                                for (int i = 0; i < FormUpravnik.Prostorije.Count; i++)
+                                {
+                                    if (FormUpravnik.Prostorije[i].BrojProstorije == row.BrojProstorije)
+                                    {
+                                        FormUpravnik.Prostorije.Remove(FormUpravnik.Prostorije[i]);
+                                    }
+                                }
+
+                                break;
+                            }
                         }
                     }
                 }
